New single experiment page no tabs (work in progress)
